backend/requirements.txt
Flask
numpy
Pillow
tensorflow
keras
